Add new RMW storage implementation
Despite it looks little bit low-level but it works well with unaligned access.
I/O operations optimized by splitting it to stages:

- Not aligned start
- Aligned middle
- Not Aligned end

Each stage may or may not be needed depending from memory offset and data
length.

Also the data checked to determine that erase and/or write needed or not.